Debug env for tox
Running a test with pdb was difficult because testr captures
output and causes pdb prompt to quit.

Tips for how to run testr with debug are provided here:
  https://wiki.openstack.org/wiki/Testr#Debugging_.28pdb.29_Tests

This change puts these commands into a debug env in tox.ini so
you can do a command like

 tox -e debug test_auth_token_middleware

or, for more granularity:
 tox -e debug keystone.tests.test_sql_migrate_extensions.EndpointFilterExtension
 tox -e debug keystone.tests.test_sql_migrate_extensions.EndpointFilterExtension.test_downgrade

and when it hits your breakpoint you'll get the debug prompt.

This same change was made to keystone in commit c8302509 .

Change-Id: I2b5dc81f4652dc89ecbcc7c785245a81cf77a1ca