Expose auth methods on the adapter
Provide access to get_token, get_endpoint and invalidate to the adapter.
The adapter is essentially created per individual client and it can be
useful to know things like the endpoint that requests will be sent to
based on the parameters that are included in the endpoint_filter.

This essentially allows us to emulate the management_url and auth_token
properties of the existing clients.

Change-Id: Ic01bc52bb38e8fb72e7a6d93bfd2944b11d0b070