don't log service catalog in every token response
The whole service catalog is embedded in every token, and by default
all token responses are logged at DEBUG. This adds a huge amount of
basically const data into system logs, over and over and over again.

We should not log the service catalog by default on every token
response. The following replaces the service catalog with the token
<removed>.

This reduces the compressed logs of API services by about 1/3.

Change-Id: I95832d0f13ca93c4618784da9d1eb9ca166cae53