Don't return default for domain in v2 accessinfo
The V2 AccessInfo returns 'default' and 'Default' for domain_id and
domain_name parameters respectively because this information is not
included in the V2 token and it was trying to help.

Unfortunately this makes it impossible to tell if there is no domain
information available or the domain information is actually default
which may not be the case on all systems.

Make it so V2AccessInfo returns None for domain properties.

Change-Id: Ieb79c70ebe8e73b7f1a445015a3eb00bda322e8a