Ensure that failing responses are logged
The boolean value of a failed response is False and so the way we
populate the log output does not work when the request failed.

When logging check that a response is not None rather than simply
checking it's boolean value.

Change-Id: I07fb46f156fdf8267fd3d4dc7c587cd604838d73
Closes-Bug: #1451625
(cherry picked from commit c7ec27a448d20201471a8f953e3b58b9ded589aa)