Remove oslo.i18n dependency
Remove the oslo.i18n dependency from keystoneauth library. This is
to limite the external dependency graph to ensure there is the minimal
surface area that can/could break due to a library update.

This change simply replaces the translation hint functions with a no-op
lambda. As keystoneauth continues to be developed the use of these no-op
functions can be removed.

Change-Id: Iec422765ab0fdf3c23fdf956be2caacff8e3386d