Use auth_type instead of auth_plugin by default
This was something that OSC has pushed for a long time now, and it's
better to be consistent than worry about wording.

Change --os-auth-plugin and auth_plugin= in CONF to be --os-auth-type
and auth_type= by default. We deprecated (but will probably keep for a
long time) the auth_plugin name.

Change-Id: I9748aeb4a490f88c73ef22ebe49e9b4ac6af78cd