Allow fetching oslo.config Opts from plugins
At the moment we only have access to the loading.Opts for a plugin,
which is fine most of the time but in places like auth_token middleware
we sometimes need to manually register a plugin options.

Extend the existing function to get the options from a Loader directly
if one is passed.

Change-Id: I3fd0cead78d5cc8ee2a3d19f056b5eff15208f84