Specify default_domain to generic plugin
The generic plugin is supposed to work with both the v2 and v3 APIs.
This doesn't necessarily work because you either need to provide domain
information or not which implies specifying a v2 or v3 preference.

By adding default domain we can allow using v2 or v3 authentication
interchangeably. This is something that openstackclient does already.

Closes-Bug: #1515041
Change-Id: I8d036a080a09b9310ffdd73d3501b6df29a212b9