Make public the base loader classes
The loaders are a companion part to the authentication plugins that
assist in loading a specific plugin from options supplied via a config
file, command line arguments or other mechanisms.

The Base loaders handle the options that are common to a specific
identity authentication mechanism. They were added to the private
directory in the move from keystoneclient as they are only used by the
loaders defined in that directory and it gave us time to settle the
interfaces.

They are not expected to be private forever as they are expected to be
reused by anyone that is developing identity plugins outside of the
keystoneauth repository.

Closes-Bug: #1507423
Change-Id: I4b2fc321ae75e92509aacb15ac21e96880f2b20e