Use XML matcher
The XML comparisons were done using strings that were slightly
normalized by removing newlines. This isn't a safe way to do XML
comparison since XML documents that are represented differently as
strings can still be the same. For example, the shortened
namespace can be different or attributes can be in different order.

XML comparisons should be done using the XML matcher so that
differences in the exact representation are allowed.

Change-Id: Id26531eb0f55b45a257ea22eb7263524158f45f9