Tweak the way plugin attributes are loaded
By using the setdefault method to add plugin parameters to the kwargs
passed to load_from_options() we are unnecessarily calling the getter
function when the value may not be saved.

This has generally not been a problem however if we start doing things
like prompting users for a password in the getter this is an expensive
and disruptive operation for a parameter that may not be stored.

Change-Id: Ic8dc3a425662c4a89372f178b175b8caab44e024