Add oslo.config compatibility attributes
In trying to get the OpenStack server projects to consume keystoneauth,
we ran in to the fact that they want to treat the things registered with
"register_conf_options" as oslo.config objects. It works fine that they
aren't - except that an oslo.config option has an attribute called
'deprecated_opts' while ours is called 'deprecated'. The same is true
of the other three attributes added.

Change-Id: I7d53c17d518b01946e091702a582648998c2ae17