Allow parameter expansion in endpoint_override
Allow %(project_id)s and %(user_id)s in an endpoint_override string to
be replaced with the appropriate values from the authentication. This
makes it easier to use and specify overrides from the command line or
scripts where we may not know the project_id/user_id ahead of time.

Change-Id: Ia09f5a88cb6590113dab26796c58e708014e98f7
Closes-Bug: #1536874