Include query string in request logging
While debugging openstacksdk, I noticed that the way we send query
parameters in a request led to them not being included in the URL that
gets logged within _http_log_request. When the `url` directly includes
the query string, it of course gets logged. When a caller utilizes the
`params` argument (set to a dictionary of query parameters), which is
subsequently handled inside of the requests library, those parameters
don't end up getting logged.

This change grabs any query parameters out of the `params` argument and
sends them into the log method. The resulting output of a url with
included query strings is wrapped in double quotes so the CURL command
that gets logged is actually runnable, because the question mark
separator is interpreted by the shell.

Change-Id: I249ae29e0d60887c637e497133c2a134f38982b5