Use betamax hooks to mask fixture results
We cannot rely on native recorded fixtures with betamax,
because of two reasons: - betamax is recording cloud usernames
and passwords. If we run this into real clouds, that's a security
problem. We need to replace those by dummy users and passwords -
when keystone tokens are generated, they are recorded with a fixed
expiration time. That means, that tests are passing until the token
expire, but after that, tests are no longer valid because the
recorded token is in the past.

So need to use pre_record hook to alter the recorded results.
The hook has to update fixtures with 2 things: - mask user,
password and tenant for clouds - set a token expiration date in
the long future.

Change-Id: I5d959ee9386fb0321e9d39c73792155acd6b4851
Closes-Bug: 1570384