oidc: Remove unused parameters in _OidcBase
Some methods in _OidcBase accepted a parameter to indicate the URL where
a request was to be made:

- _get_keystone_token accepted federated_token_url, the URL for
  federated authentication.
- _get_access_token accepted access_token_endpoint, the URL used to get
  an oidc access token.

However, the methods were using internally the object attribute
self.access_token_endpoint and the property self.federated_token_url,
therefore the parameters were completely ignored. This changes removes
them and adjust the derived classes.

Change-Id: Ibc2157d47cb0edd59c4d966bae77c102423f97ae