oidc: add OidcAccessToken class to authenticate reusing an access token
Some services or users may have obtained an access token, so it would be
possible to authenticate using this token directly (for example a
service where the user has already logged in). This new class makes
possible to use an access token to authenticate directly with Keystone,
exchanging it for a Keystone token.

Closes-bug: 1583780
Change-Id: I5a31270194a3d1aa48de709dba49afde460731e2