Add available flag to plugin loaders
Currently when iterating through plugins all plugins are loaded and
returned to the user. This is confusing for things like the kerberos
plugin where the required dependencies may not be available.

Add an available property on plugin loaders. Plugins that do not wish to
be shown to users can set available to false.

The lack of tests on this patch is unfortunate however any testing
involes a lot of mocking at levels lower than keystoneauth interact with
(i've tried). We would need to mock the pkg_resources layer that
stevedore uses and are essentially testing that EnabledExtensionManager
is doing the right thing.

I encourage people to verify this manually.

Closes-Bug: #1589740
Change-Id: I446441467ef32f7bc916221991388ac528df48f4