Fix kerberos available property
Kerberos is available if the requests_kerberos module was successfully
loaded in keystoneauth.extras.kerberos. kerberos_requests has no
meaning and causes keystoneauth1 to raise an error:

 AttributeError: "'module' object has no attribute 'kerberos_requests'"

But this gets eaten by stevedore[1]so if logging is not set up the
result is that v3kerberos silently fails to load. This patch fixes the
typo that causes this.

[1] http://git.openstack.org/cgit/openstack/stevedore/tree/stevedore/extension.py?id=01b09a5c1d683007d86fec09594db9f5787805f3#n169

Change-Id: Ie7445e99d710bd238db7ce04a8610bd54c2d6ae7