Add pretty serializer for betamax fixture
Saving json responses all on one line escaped inside of json
cassettes is great for computers, but is impossible for humans to
read. Add a serializer that is nicely flowed yaml that emits
multi-line values as yaml blocks. Additionally, re-flow and indent
the nested json, which will stay as json.

An example of the output produced can be seen at:

  https://review.openstack.org/#/c/328338/2/shade/tests/unit/fixtures/test_create_flavor.yaml

Hook it in to the keystoneauth1 betamax fixture by default, because
why in the world would you want ugly when you can have pretty.

Change-Id: I457408fcbbdca240090228d18f0482f958a7d6e4