Add a service token wrapper
There are cases from a service where you have to wrap and pass a User
and Service token together to make things work. Add a wrapper that adds
both the user and service token to requests.

This will be something we will handle differently in auth_token
middleware but should still provide for other implementations.

Change-Id: I284f799d1f9e8d33ff032376af02b64cd6bbf510