Don't use private testtools.test module
The testtools.test module is the unit testing module for testtools and
is an internal module with additional requirements. We should not be
relying on this module.

Copy the tests that were exposed from testtools into here.

Change-Id: I1e09228cff7a0c8136447f07df6864045a6fb849
(cherry picked from commit 31b9d125b0a3142ac17e8552911bb507633f05f7)