Shift additional_user_agent in the stack
The additional_user_agent is intended for things, such as
os-client-config, that sit somewhere in the stack between keystoneauth
and the actual client. So according to jamielennox in irc, it should be:

  app client *additional_user_agent keystoneauth requests

The current logic puts additional_user_agent between app and client,
leading to things that look like this:

  "User-Agent: v.py/1 os-client-config/1.26.1 shade/1.19.1
   keystoneauth1/2.18.0 python-requests/2.13.0 CPython/2.7.12"

Which is a bit off.

Change-Id: Iddfbeab322a39d3ba893a15aabb4c79050526451