Port the missing version data discovery tests from ksc
There are a number of fixtures still present in test_discovery for
testing that version discovery works with the different discovery pages
found in glance, cinder and others.

The method names for a lot of this stuff was slightly changed from
keystoneclient to keystoneauth so it looks like these tests were dropped
in the port rather than fixed.

Re-add these tests to ensure other discovery formats are supported.

Change-Id: I311e94437545544917b027de69a89ede9d96f555